Refactored the form to a seperate file and made vitamins an optional category
