+++ frontend/src/pages/AddProduct/index.tsx
@@ -2,16 +2,15 @@ import React from "react";
 import {
   Heading,
   Box,
-  Form,
   FormField,
   Text,
   TextInput,
-  Button,
   MaskedInput,
 } from "grommet";
 import { Product } from "../../util/schema/product";
 import { Redirect } from "react-router-dom";
 import { postProduct, postCSVProducts } from "../../util/data/requests";
+import ProductForm from '../../containers/ProductForm';
 
 interface Category {
   name: string;
@@ -31,199 +30,6 @@ interface NutritionField {
   unit?: "g" | "mg";
 }
 
-const formStructure: Array<Category> = [
-  {
-    name: "Product Information",
-    fields: [
-      {
-        name: "name",
-        fieldType: "text",
-        required: true,
-      },
-      {
-        name: "manufacturer",
-        fieldType: "text",
-        required: true,
-      },
-    ],
-  },
-  {
-    name: "Energy",
-    fields: [
-      {
-        name: "kj",
-        fieldType: "masked",
-        required: true,
-      },
-      {
-        name: "kcal",
-        fieldType: "masked",
-        required: true,
-      },
-    ],
-  },
-  {
-    name: "Carbohydrates",
-    fields: [
-      {
-        name: "carbohydrates",
-        fieldType: "masked",
-        required: true,
-        unit: "g"
-      },
-      {
-        name: "sugar",
-        fieldType: "masked",
-        required: true,
-        unit: "g"
-      },
-      {
-        name: "addedSugar",
-        fieldType: "masked",
-        required: true,
-        unit: "g"
-      },
-      {
-        name: "fiber",
-        fieldType: "masked",
-        required: true,
-        unit: "g"
-      },
-      {
-        name: "starch",
-        fieldType: "masked",
-        required: true,
-        unit: "g"
-      },
-    ],
-  },
-
-  {
-    name: "Fats",
-    fields: [
-      {
-        name: "fat",
-        fieldType: "masked",
-        required: true,
-        unit: "g"
-      },
-      {
-        name: "saturated",
-        fieldType: "masked",
-        required: true,
-        unit: "g"
-      },
-      {
-        name: "monosaturated",
-        fieldType: "masked",
-        required: true,
-        unit: "g"
-      },
-      {
-        name: "trans",
-        fieldType: "masked",
-        required: true,
-        unit: "g"
-      },
-    ],
-  },
-  {
-    name: "Protein",
-    fields: [
-      {
-        name: "protein",
-        fieldType: "masked",
-        required: true,
-        unit: "g"
-      },
-    ],
-  },
-  {
-    name: "Salt",
-    fields: [
-      {
-        name: "salt",
-        fieldType: "maskHighPrecision",
-        required: true,
-        unit: "g"
-      },
-    ],
-  },
-  {
-    name: "Vitamins",
-    fields: [
-      {
-        name: "a",
-        fieldType: "maskHighPrecision",
-        unit: "mg"
-      },
-      {
-        name: "d",
-        fieldType: "maskHighPrecision",
-        unit: "mg"
-      },
-      {
-        name: "e",
-        fieldType: "maskHighPrecision",
-        unit: "mg"
-      },
-      {
-        name: "k",
-        fieldType: "maskHighPrecision",
-        unit: "mg"
-      },
-      {
-        name: "b1",
-        fieldType: "maskHighPrecision",
-        unit: "mg"
-      },
-      {
-        name: "b2",
-        fieldType: "maskHighPrecision",
-        unit: "mg"
-      },
-      {
-        name: "b3",
-        fieldType: "maskHighPrecision",
-        unit: "mg"
-      },
-      {
-        name: "b5",
-        fieldType: "maskHighPrecision",
-        unit: "mg"
-      },
-      {
-        name: "b6",
-        fieldType: "maskHighPrecision",
-        unit: "mg"
-      },
-      {
-        name: "b7",
-        fieldType: "maskHighPrecision",
-        unit: "mg"
-      },
-      {
-        name: "b9",
-        fieldType: "maskHighPrecision",
-        unit: "mg"
-      },
-      {
-        name: "b12",
-        fieldType: "maskHighPrecision",
-        unit: "mg"
-      },
-      {
-        name: "c",
-        fieldType: "maskHighPrecision",
-        unit: "mg"
-      },
-    ],
-  },
-];
-
-const capitalise = (value: string) => {
-  return value.charAt(0).toUpperCase() + value.slice(1);
-};
 
 const propertyToNumber = (property: number): number => {
   if (property) {
@@ -301,83 +107,16 @@ const fileChosen = (file: any | undefined, setReport: any) => {
   reader.readAsText(file);
 };
 
-const mapFields = (fields: Array<NutritionField>) => {
-  return fields.map((field: NutritionField) => {
-    const label = `${capitalise(field.name)} ${field.unit ? `(${field.unit})` : ""} ${field.required ? " *" : ""}`
-
-    if (field.fieldType === "maskHighPrecision") {
-      return (
-        <FormField
-          name={field.name}
-          label={label}
-          required={field.required || false}
-        >
-          <MaskedInput
-            name={field.name}
-            mask={[
-              {
-                regexp: /^\d+$/,
-                placeholder: "0",
-              },
-              { fixed: "." },
-              {
-                length: 3,
-                regexp: /^[0-9]{1,4}$/,
-                placeholder: "000",
-              },
-            ]}
-          />
-        </FormField>
-      );
-    } else if (field.fieldType === "masked") {
-      return (
-        <FormField
-          name={field.name}
-          label={label}
-          required={field.required || false}
-        >
-          <MaskedInput
-            name={field.name}
-            mask={[
-              {
-                regexp: /^\d+$/,
-                placeholder: "0",
-              },
-              { fixed: "." },
-              {
-                length: 2,
-                regexp: /^[0-9]{1,4}$/,
-                placeholder: "00",
-              },
-            ]}
-          />
-        </FormField>
-      );
-    }
-
-    return (
-      <FormField
-        name={field.name}
-        label={label}
-        required={field.required || false}
-      >
-        <TextInput name={field.name} />
-      </FormField>
-    );
-  });
-};
-
-const mapCategories = (categories: Array<Category>) => {
-  return categories.map((c: Category) => (
-    <>
-      <Heading level={2}>{c.name}</Heading>
-      {mapFields(c.fields)}
-    </>
-  ));
-};
 
 const AddProduct = () => {
   const [sent, setSent] = React.useState(false);
+
+  const onSubmit = (event: any) => {
+    const product = toProduct(event.value);
+    postProduct(product).then((result) => {
+      setSent(true);
+    });
+  };
 
   return (
     <Box pad="large" align="center">
@@ -385,21 +124,7 @@ const AddProduct = () => {
         <Heading>Add Product</Heading>
         <Text>All values should be provided per 100 g / ml</Text>
         <Box width="large">
-          <Form
-            onSubmit={(event: any) => {
-              const product = toProduct(event.value);
-              postProduct(product).then((result) => {
-                setSent(true);
-              });
-            }}
-          >
-            {mapCategories(formStructure)}
-
-            <Box direction="row" gap="medium">
-              <Button type="submit" primary label="Submit" />
-              <Button type="reset" label="Reset" />
-            </Box>
-          </Form>
+          <ProductForm onSubmit={onSubmit} />
         </Box>
         {sent && <Redirect to="/products" />}
       </Box>
